+++ packages/ui-positions/src/components/rwaStatus/rwaStatus.tsx
@@ -1,6 +1,6 @@
 import { useGetRwaSymbolOpenStatus } from "@veltodefi/hooks";
 import { useTranslation } from "@veltodefi/i18n";
-import { cn, Flex, Text } from "@veltodefi/ui";
+import { Box, Tips, cn, Text } from "@veltodefi/ui";
 
 export const RwaStatusTag = ({ symbol }: { symbol: string }) => {
   const { isRwa, open } = useGetRwaSymbolOpenStatus(symbol);
@@ -11,20 +11,25 @@ export const RwaStatusTag = ({ symbol }: { symbol: string }) => {
   }
 
   return (
-    <Flex
-      r="base"
-      px={2}
-      height={18}
-      className={cn(
-        "oui-shrink-0",
-        open ? "oui-bg-success/15" : "oui-bg-danger/15",
-      )}
-    >
-      <Text size="2xs" color={open ? "success" : "danger"}>
-        {open
-          ? t("trading.rwa.marketHours")
-          : t("trading.rwa.outsideMarketHours")}
-      </Text>
-    </Flex>
+    <Tips
+      title={t("common.tips")}
+      content={
+        <Text color={open ? "success" : "danger"}>
+          {open
+            ? t("trading.rwa.marketHours")
+            : t("trading.rwa.outsideMarketHours")}
+        </Text>
+      }
+      trigger={
+        <Box p={2} className="oui-flex oui-cursor-pointer oui-items-center">
+          <Box
+            width={4}
+            height={4}
+            r="full"
+            className={cn(open ? "oui-bg-success" : "oui-bg-danger")}
+          />
+        </Box>
+      }
+    />
   );
 };