+++ packages/net/CHANGELOG.md
@@ -478,6 +478,84 @@
 
 - publish
 
+## 2.12.4
+
+### Patch Changes
+
+- publish
+- 07c88a0: publish
+
+## 2.12.4-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.12.3
+
+### Patch Changes
+
+- publish
+- 528e3f4: publish
+
+## 2.12.3-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.12.2
+
+### Patch Changes
+
+- publish
+- 06007f7: publish
+
+## 2.12.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.12.1
+
+### Patch Changes
+
+- publish
+- 75b87bc: publish
+
+## 2.12.1-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.12.0
+
+### Minor Changes
+
+- 1053a3d: publish
+- publish
+
+## 2.12.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+## 2.11.2
+
+### Patch Changes
+
+- publish
+- 1c40427: publish
+
+## 2.11.2-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.11.1
 
 ### Patch Changes